# modules/analytics/job-submission-service/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Flink Job Submission API",
        "description": "API for submitting, monitoring, and managing Flink jobs",
        "version": "1.0.0",
        "contact": {
            "name": "IoT Platform Team"
        }
    },
    "servers": [
        {
            "url": "/"
        }
    ],
    "paths": {
...
        "/jobs/{jobId}": {
            "get": {
                "summary": "Get job",
                "description": "Gets details for a specific job",
                "parameters": [
                    {
                        "name": "jobId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "Job ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}